add MPI as extension
